Fixed Seg Fault and added Functionality
The Program will now compile and Do something...
It is capable of drawing a maze like grid but the walls are not quite right.
The edges are always filled in and the bottom row always seems to be filled in.
The rest of the walls seem to be filled in at random.

TODO
Fix this. Almost done.